db: Delete dependant 'links' before 'vl_entries'
When we 'fabsys db-clean' a brun, or delete an empty brun after
'fabsys dump-delete', we try to delete some vl_entries in the brun.
If there are 'links' entries that reference a vl_entries entry that
we're trying to delete, the delete will fail due to foreign key
constraints, since 'links' has a foreign key constraint on
vl_entries.id.

Normally, this shouldn't happen, since we only do this when there is
no volume_dump for the associated vl_entry. And if there's no
volume_dump, we shouldn't have any links recorded, since we never
recorded the dump for the volume. But this can happen in some edge
cases if dump processing is interrupted after the links are recorded.

So to avoid these errors, before deleting vl_entries, delete any
referencing 'links' entries.

Add a 'links' entry in our tests sample db that references a
vl_entries entry with no corresponding volume_dump, to test this case.