rpm: Add sqlite to BuildRequires on EL8/Fedora
For our tests to run, we need the 'sqlite3' command in addition to the
python library. On RHEL7, sqlite3 is always installed (yum requires
it), but this isn't guaranteed on RHEL8 and Fedora. Building our RPM
runs the tests by default, so building RPMs on RHEL8 and Fedora breaks
if sqlite3 isn't installed.

To fix building RPMs on RHEL8 and Fedora, explicitly add a
BuildRequires on sqlite, so the sqlite3 command is available when the
tests run.