Reject localauth option in fs_auth()
When the localauth option is enabled, any FABS operation that calls
fs_auth() fails with the error:

    fs: Unrecognized or ambiguous switch '-localauth'

This is because the underlying fs command suite, unlike other AFS
commands, lacks a native -localauth option.

This error message can be confusing. To address this, a check is now
performed within fs_auth() to determine if localauth is enabled before
executing an fs command. If it is, a clear and actionable exception is
raised that explains the incompatibility and suggests alternative
authentication methods. The error message that is logged when the
exception is raised looks like this:

    fabs.restore.cfg_err ERROR: Error when processing restore request 41
    Traceback (most recent call last):
      ...
    fabs.err.ConfigError: Cannot run 'fs' commands with localauth. FABS
    currently provides limited support for the afs/localauth option, and
    its use is not recommended in production environments. To avoid this
    error, use another authentication method.

This check is specific to fs commands and does not affect other command
suites (like vos) that support localauth.